- Checkpoint

+++ lib/LcdBarGraph/LcdBarGraph.hpp
@@ -9,7 +9,8 @@ public:
     LcdBarGraph(LCD* lcd, uint8_t startSlot, uint8_t middleSlot, uint8_t endSlot);
 
     auto init() -> void;
-    auto draw(uint8_t slot, uint8_t posCol, uint8_t posRow, uint8_t barLength, double percent) -> void;
+    auto draw(uint8_t slot, int16_t posCol, int16_t posRow, int16_t barLength, double percent) -> void;
+
 private:
     LCD *lcd;
     uint8_t startSlot, middleSlot, endSlot;
